# API-Go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.Body:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  user.CreateUserRequest:
    properties:
      actor_id:
        type: string
      actor_type:
        enum:
        - USER
        - INSTRUCTOR
        - LEARNER
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - actor_type
    - email
    - first_name
    - last_name
    - password
    type: object
  user.ListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/user.UserResponse'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  user.UpdateUserRequest:
    properties:
      first_name:
        type: string
      is_active:
        type: boolean
      last_name:
        type: string
    type: object
  user.UserResponse:
    properties:
      actor_type:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      last_name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'API REST de Identity and Access Management (IAM): autenticación JWT,
    RBAC, usuarios, roles, funcionalidades, sesiones y auditoría.'
  title: API JeussAirel (IAM)
  version: "1.0"
paths:
...
  /sessions/{id}:
    delete:
      description: Revoca una sesión del usuario autenticado por su ID.
      parameters:
      - description: UUID de la sesión
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sesión revocada
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Token inválido o expirado
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Sesión no encontrada
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Revocar sesión
      tags:
      - Sessions
  /sessions/active:
    get:
      description: Devuelve las sesiones (refresh tokens) activas del usuario autenticado.
      responses:
        "200":
          description: Sesiones activas
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: Token inválido o expirado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar sesiones activas
      tags:
      - Sessions
  /users:
    get:
      consumes:
      - application/json
      description: Devuelve una lista paginada de usuarios.
      parameters:
      - description: Número de página (default 1)
        in: query
        name: page
        type: integer
      - description: Tamaño de página 1-100 (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuarios obtenidos
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/user.ListResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_USER_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Listar usuarios
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Registra un nuevo usuario en el sistema.
      parameters:
      - description: Datos del usuario
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Usuario creado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "400":
          description: Datos inválidos
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_USER_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Crear usuario
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina (desactiva) un usuario por su UUID.
      parameters:
      - description: UUID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuario eliminado
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: Sin permisos (IDENTITY_USER_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Eliminar usuario
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Devuelve un usuario por su UUID.
      parameters:
      - description: UUID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuario encontrado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_USER_VIEW:read)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Obtener usuario por ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Actualiza los campos editables de un usuario.
      parameters:
      - description: UUID del usuario
        in: path
        name: id
        required: true
        type: string
      - description: Campos a actualizar
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Usuario actualizado
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/user.UserResponse'
              type: object
        "403":
          description: Sin permisos (IDENTITY_USER_MANAGE:write)
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: Usuario no encontrado
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Actualizar usuario
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: 'Token JWT de acceso. Uso: `Authorization: Bearer <access_token>`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"